% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_probabilities.R
\name{allocate_probabilities}
\alias{allocate_probabilities}
\title{Reallocate the treatment randomization probabilities via Thompson Sampling}
\usage{
allocate_probabilities(posterior, c, optimize = "max")
}
\arguments{
\item{posterior}{stanreg object that contains the posterior samples}

\item{c}{float between 0 and 1 indicating how much to stablize probabilies}

\item{optimize}{character indicating whether to optimize min or max}
}
\value{
Dataframe containing the treatment allocation probabilities by id
}
\description{
Reallocate the treatment randomization probabilities via Thompson Sampling
}
