% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin}
\alias{burnin}
\title{Function for simulating the burn-in phase of the Platform-of-1 design}
\usage{
burnin(n_trts, n_burn_cycles, burn_obvs_per_period, betas, y_sigma, phi)
}
\arguments{
\item{n_trts}{(integer): The number of treatments. The value of n_trts should be a positive integer.}

\item{n_burn_cycles}{(integer): The number of burn-in cycles. This parameter defines how many times the dataset will be replicated for the burn-in phase. The value of n_burn_cycles should be a positive integer.}

\item{burn_obvs_per_period}{(integer): The number of observations per period for each treatment during the burn-in period. The value of burn_obvs_per_period should be a positive integer.}

\item{betas}{(numeric vector): A vector of regression coefficients representing the effects of the treatments. The length of this vector should be equal to n_trts.}

\item{y_sigma}{(numeric): The standard deviation of the error term in the outcome variable. This parameter represents the noise in the outcome variable and should be a positive number.}

\item{phi}{(numeric): The autoregressive (AR) coefficient for the outcome variable. If phi is 0, there is no serial correlation in the outcome variable. Otherwise, the outcome variable is generated with an AR(1) process using the specified phi value.}
}
\value{
A tibble containing treatment indicators for all treatments and an outcome
}
\description{
Function for simulating the burn-in phase of the Platform-of-1 design
}
