% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IND.R
\name{IND}
\alias{IND}
\title{Title}
\usage{
IND(
  n_subj,
  n_trts,
  n_periods,
  n_obvs,
  betas,
  y_sigma,
  chains,
  warmup,
  iter,
  adapt_delta = 0.99,
  max_treedepth = 15
)
}
\arguments{
\item{max_treedepth}{}
}
\value{

}
\description{
Title
}
