% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Wrapper for running rstan}
\usage{
mcmc(
  file,
  data,
  chains = 1,
  warmup = 1000,
  iter = 3000,
  adapt_delta = 0.99,
  max_treedepth = 15
)
}
\arguments{
\item{data}{List of data from the trial}

\item{max_treedepth}{}
}
\value{

}
\description{
Wrapper for running rstan
}
