% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilize_probs.R
\name{stabilize_probs}
\alias{stabilize_probs}
\title{Function for stabilizing allocation probabilities}
\usage{
stabilize_probs(probs, c)
}
\arguments{
\item{c}{}
}
\value{

}
\description{
Function for stabilizing allocation probabilities
}
