% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Function for converting Stan-friendly data to be more tidyverse friendly}
\usage{
tidy_data(standata)
}
\arguments{
\item{standata}{List containing the data that was}

\item{n_obvs}{not needed?}
}
\value{

}
\description{
Function for converting Stan-friendly data to be more tidyverse friendly
}
