% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_probabilities.R
\name{allocate_probabilities}
\alias{allocate_probabilities}
\title{Reallocate the treatment randomization probabilities via Thompson Sampling}
\usage{
allocate_probabilities(posterior, n_trts, optimize = "max")
}
\arguments{
\item{posterior}{stanreg object that contains the posterior samples}

\item{n_trts}{integer indicating number of treatments currently in trial}

\item{optimize}{character indicating if the maximum or minimum reward should be optimized for}
}
\value{
Dataframe containing the treatment allocation probabilities by id
}
\description{
Reallocate the treatment randomization probabilities via Thompson Sampling
}
