% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Title}
\usage{
simulate(
  n_trts,
  n_burn_cycles,
  burn_obvs_per_period,
  adaptive_obvs_per_period,
  max_duration,
  betas,
  y_sigma,
  priors,
  n_chains,
  n_iter,
  phi,
  stabilize = NULL,
  objective = "Maximize",
  adapt_delta = 0.999,
  max_treedepth = 15
)
}
\arguments{
\item{max_treedepth}{}
}
\description{
Title
}
